# backend/api/openapi.yaml
openapi: "3.1.0"
info:
  title: SpendWise API
  description: Collaborative household financial ledger API
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/workspaces/{workspaceId}/transactions:
    post:
      operationId: createTransaction
      tags: [Transactions]
      summary: Create a transaction with entries (editor+ required)
      description: |
        Creates a transaction with one or more entries. Each entry assigns an amount
        to a category and optionally a participant. Entry amounts MUST sum to the
        transaction total.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    get:
      operationId: listTransactions
      tags: [Transactions]
      summary: List transactions in date range (viewer+ required)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
            example: "2026-02-01"
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
            example: "2026-03-01"
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransactionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/workspaces/{workspaceId}/transactions/import:
    post:
      operationId: importTransactions
      tags: [Transactions]
      summary: Import transactions from CSV (editor+ required)
      description: |
        Imports transactions from a CSV file. Duplicate fingerprints are skipped.
        Returns the list of created transactions.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        "200":
          description: Import results
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: integer
                    description: Number of transactions imported
                  skipped:
                    type: integer
                    description: Number of duplicates skipped
                  transactions:
                    type: array
                    items:
                      $ref: "#/components/schemas/TransactionResponse"
                required: [imported, skipped, transactions]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/workspaces/{workspaceId}/transactions/{transactionId}:
    get:
      operationId: getTransaction
      tags: [Transactions]
      summary: Get transaction with entries (viewer+ required)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/TransactionId"
      responses:
        "200":
          description: Transaction details with entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteTransaction
      tags: [Transactions]
      summary: Delete a transaction (editor+ required)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/TransactionId"
      responses:
        "204":
          description: Transaction deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ──────────────────────────────────────────
  # Funding
  # ──────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtained from `/api/v1/auth/{provider}/callback`
  parameters:
...
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    TransactionId:
      name: transactionId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "workspace not found"
      required: [error]
    Money:
      type: object
      description: Monetary amount with currency
      properties:
        amount:
          type: string
          description: Decimal amount as string for precision
          example: "150.75"
        currency:
          type: string
          default: PLN
          example: PLN
      required: [amount, currency]

    # ── Auth ──
...
    CreateTransactionRequest:
      type: object
      properties:
        description:
          type: string
          example: Weekly groceries at Biedronka
        date:
          type: string
          format: date
          example: "2026-02-12"
        total_amount:
          $ref: "#/components/schemas/Money"
        notes:
          type: string
          default: ""
        entries:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CreateEntryRequest"
      required: [description, date, total_amount, entries]
    CreateEntryRequest:
      type: object
      properties:
        category_id:
          type: integer
          format: int64
        participant_id:
          type: integer
          format: int64
          nullable: true
          description: User ID of the participant (null = unassigned)
        amount:
          $ref: "#/components/schemas/Money"
        note:
          type: string
          default: ""
      required: [category_id, amount]
    TransactionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        workspace_id:
          type: integer
          format: int64
        created_by:
          type: integer
          format: int64
        description:
          type: string
        date:
          type: string
          format: date
        total_amount:
          $ref: "#/components/schemas/Money"
        source:
          type: string
          enum: [manual, import]
        notes:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/EntryResponse"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, workspace_id, created_by, description, date, total_amount, source, entries, created_at, updated_at]
    EntryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        category_id:
          type: integer
          format: int64
        participant_id:
          type: integer
          format: int64
          nullable: true
        amount:
          $ref: "#/components/schemas/Money"
        note:
          type: string
      required: [id, category_id, amount]

    # ── Funding ──
...